+++ plugins/woocommerce-admin/packages/components/src/card/index.js
@@ -18,16 +18,24 @@ import { validateComponent } from '../lib/proptype-validator';
  */
 class Card extends Component {
 	render() {
-		const { action, children, menu, title } = this.props;
+		const { action, children, description, isInactive, menu, title } = this.props;
 		const className = classnames( 'woocommerce-card', this.props.className, {
 			'has-menu': !! menu,
 			'has-action': !! action,
+			'is-inactive': !! isInactive,
 		} );
 		return (
 			<div className={ className }>
 				{ title && (
 					<div className="woocommerce-card__header">
-						<H className="woocommerce-card__title woocommerce-card__header-item">{ title }</H>
+						<div className="woocommerce-card__title-wrapper">
+							<H className="woocommerce-card__title woocommerce-card__header-item">{ title }</H>
+							{ description && (
+								<H className="woocommerce-card__description woocommerce-card__header-item">
+									{ description }
+								</H>
+							) }
+						</div>
 						{ action && (
 							<div className="woocommerce-card__action woocommerce-card__header-item">
 								{ action }
@@ -53,6 +61,14 @@ Card.propTypes = {
 	 * Additional CSS classes.
 	 */
 	className: PropTypes.string,
+	/**
+	 * The description displayed beneath the title.
+	 */
+	description: PropTypes.oneOfType( [ PropTypes.string, PropTypes.node ] ),
+	/**
+	 * Boolean representing whether the card is inactive or not.
+	 */
+	isInactive: PropTypes.bool,
 	/**
 	 * An `EllipsisMenu`, with filters used to control the content visible in this card
 	 */